the floating block is haunted.

+++ src/main/java/teamport/aether/entity/EntityFloatingBlock.java
@@ -21,7 +21,9 @@ import teamport.aether.blocks.BlockLogicOreGravitite;
 ///  all the comment were made to understand what the logic does
 public class EntityFloatingBlock extends EntityFallingBlock {
     public boolean onCeiling = false;
+    public int ceilingTime = 0;
     public final int maxFloatingTime = 600;
+    public boolean pushedByPiston = false;
 
     public EntityFloatingBlock(World world) {
         super(world);
@@ -48,6 +50,7 @@ public class EntityFloatingBlock extends EntityFallingBlock {
             this.remove();
             return;
         }
+
         //--------------------------------------------------------------------------------------------------------------
         // move the block do some calc
         this.pushesThisTick = 0;
@@ -55,16 +58,24 @@ public class EntityFloatingBlock extends EntityFallingBlock {
         if (this.pushTime < 0.05F || (double) this.pushTime < 0.25 && this.onGround) {
             this.pushTime = 0.0F;
         }
+
+        // multiplying by -1 results in the block rising forever
         int x = MathHelper.floor(this.x - 0.5);
-        int y = MathHelper.floor(this.y); // multiplying by -1 results in the block rising forever
+        int y = MathHelper.floor(this.y);
         int z = MathHelper.floor(this.z - 0.5);
         this.isOnCeiling(x, y, z);
+
+        if (onCeiling) ceilingTime++;
+        else ceilingTime = 0;
 
         this.xo = this.x;
         this.yo = this.y;
         this.zo = this.z;
         ++this.fallTime;
-        this.yd += 0.04; // rising block
+
+        // rising block
+        this.yd += 0.04;
+
         this.move(this.xd, this.yd, this.zd);
         this.xd *= 0.98;
         this.yd *= 0.98; // does the sand do that too? hm no sand multiplies what does this do?
@@ -72,37 +83,36 @@ public class EntityFloatingBlock extends EntityFallingBlock {
 
 
         //--------------------------------------------------------------------------------------------------------------
-
         if (this.world.getBlockId(x, y, z) == this.carriedBlock.blockId && !this.hasRemovedBlock) {
             this.world.setBlockWithNotify(x, y, z, 0);
             this.hasRemovedBlock = true;
         }
+
         //--------------------------------------------------------------------------------------------------------------
         // idk when this is getting hit, it seems once the block is moved its never onGround
         if (this.onGround) {
             this.xd *= 0.7;
             this.zd /= 0.7;
             this.yd *= 0.5;
-            // so the entity is removed but why? when its this hits in sand?
-            this.remove();
         }
 
         if (this.onCeiling || y > 600) {
             this.xd *= 0.7;
             this.zd /= 0.7;
             this.yd *= 0.5;
-            this.remove();
+
+            // JANK JANK JANK JANK JANK
+            if (pushedByPiston && ceilingTime < 60) {}
+            else this.remove();
         }
+
         //--------------------------------------------------------------------------------------------------------------
         // what does this do?
-        double v = Math.hypot(this.xd, this.zd);
-        if (v < 0.001 || this.isInWall()) {
-//            if (!this.onGround && !this.isInWall()) {
+        double speed = Math.hypot(this.xd, this.zd);
+        if (speed < 0.001 || this.isInWall()) {
+
             if (!this.onCeiling && !this.isInWall()) {
                 if (this.fallTime > maxFloatingTime && !this.world.isClientSide) {
-                    if (this.hasRemovedBlock) {
-                        this.drop();
-                    }
                     this.drop();
                     this.ejectRider();
                     this.remove();
@@ -110,49 +120,46 @@ public class EntityFloatingBlock extends EntityFallingBlock {
                 this.carriedBlock.heldTick(this.world, this);
                 return;
             }
+
             //----------------------------------------------------------------------------------------------------------
             // steel piston up here
             Entity rider = this.getPassenger();
             this.ejectRider();
-            this.remove();
             TileEntity oldEnt;
             //
-            if (
-                (
-                    // can pass though block that allow placement on the bottom
-                    !this.world.canBlockBePlacedAt(this.carriedBlock.blockId, x, y, z, true, Side.BOTTOM)
-                    || BlockLogicOreGravitite.canFallAbove(this.world, x, y + 1, z)
-                    // we check and place the block here so this should at some point return true and place
-                    || !this.world.setBlockWithNotify(x, y, z, this.carriedBlock.blockId)
-                )
-                && !this.world.isClientSide
-            ) {
 
-                if (this.hasRemovedBlock) {
-                    this.drop();
+            if (!this.world.isClientSide) {
+                if (world.canBlockBePlacedAt(this.carriedBlock.blockId, x, y, z, true, Side.BOTTOM)
+                        && !BlockLogicOreGravitite.canFallAbove(this.world, x, y + 1, z)
+                ) {
+
+                    boolean blockPlacedSuccessfully = this.world.setBlockWithNotify(x, y, z, this.carriedBlock.blockId);
+                    if (!blockPlacedSuccessfully) this.drop();
                 }
-            }
-            // this seems to be the server side
-            else if (!this.world.isClientSide) {
-                this.world.setBlockMetadataWithNotify(x, y, z, this.carriedBlock.metadata);
-                // this seems more concerned with rider than actual block setting logic
-                if (this.carriedBlock.entity != null) {
-                    oldEnt = this.world.getTileEntity(x, y, z);
-                    if (oldEnt != null) {
-                        oldEnt.invalidate();
+
+                else {
+                    this.world.setBlockMetadataWithNotify(x, y, z, this.carriedBlock.metadata);
+
+                    // this seems more concerned with rider than actual block setting logic
+                    if (this.carriedBlock.entity != null) {
+                        oldEnt = this.world.getTileEntity(x, y, z);
+                        if (oldEnt != null) oldEnt.invalidate();
+
+                        // not sure what happens here
+                        this.carriedBlock.entity.validate();
+                        this.carriedBlock.entity.x = x;
+                        this.carriedBlock.entity.y = y;
+                        this.carriedBlock.entity.z = z;
+                        this.carriedBlock.entity.worldObj = this.world;
+                        this.carriedBlock.entity.carriedBlock = null;
+                        this.world.replaceBlockTileEntity(x, y, z, this.carriedBlock.entity);
                     }
-                    // not sure what happens here
-                    this.carriedBlock.entity.validate();
-                    this.carriedBlock.entity.x = x;
-                    this.carriedBlock.entity.y = y;
-                    this.carriedBlock.entity.z = z;
-                    this.carriedBlock.entity.worldObj = this.world;
-                    this.carriedBlock.entity.carriedBlock = null;
-                    this.world.replaceBlockTileEntity(x, y, z, this.carriedBlock.entity);
+
+                    // notify surrounding blocks
+                    this.world.notifyBlockChange(x, y, z, this.carriedBlock.blockId);
                 }
-                // notify surrounding blocks
-                this.world.notifyBlockChange(x, y, z, this.carriedBlock.blockId);
             }
+
             //----------------------------------------------------------------------------------------------------------
             // rider stuff
             if (rider != null) {
@@ -161,6 +168,7 @@ public class EntityFloatingBlock extends EntityFallingBlock {
                     rider.startRiding((IVehicle) oldEnt);
                 }
             }
+
             //----------------------------------------------------------------------------------------------------------
             // ticks again?
             this.carriedBlock.heldTick(this.world, this);
